+++ HISTORY.md
@@ -1,5 +1,6 @@
 # Update 1.1.0 (Not out yet)
 
 - Added BTR_BAList_toArray method.
+- Added BTR_BLList_toArray method.
 
 - Fixed many BTR_NO_PREFIX mistakes.